chore(*): Move sizing test function the the correct file within the testUtils
