+++ projects/igniteui-angular/src/lib/core/utils.ts
@@ -638,10 +638,3 @@ export function getComponentCssSizeVar(size: string) {
             return 'var(--ig-size, var(--ig-size-large))';
     }
 }
-
-/**
- * Sets element size as a inline style
- */
-export function setElementSize(element: HTMLElement, size: string) {
-    element.style.setProperty('--ig-size', size);
-}
